refactor(consensus): remove types.h and add common.h
