+++ src/noir/consensus/common.h
@@ -1,0 +1,12 @@
+// This file is part of NOIR.
+//
+// Copyright (c) 2022 Haderech Pte. Ltd.
+// SPDX-License-Identifier: AGPL-3.0-or-later
+//
+#pragma once
+#include <noir/common/helper/rust.h>
+#include <noir/common/log.h>
+#include <noir/common/time.h>
+#include <noir/common/types.h>
+#include <noir/crypto/hash.h>
+#include <tendermint/types/types.pb.h>
