+++ src/noir/consensus/evidence/evidence_pool.h
@@ -4,7 +4,7 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 //
 #pragma once
-#include <noir/consensus/types.h>
+#include <noir/consensus/common.h>
 
 namespace noir::consensus::evidence {
